pyproject.toml
[tool.poetry]
name = "utmb-enrich"
version = "0.0.1"
description = ""
authors = ["Me myself and I <notmymail@bar.com>"]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"

emoji-country-flag = ">=1.3.2"
httpx = ">=0.23.3"
loguru = ">=0.6.0"
numpy = ">=1.24.2"
pandas = ">=1.5.3"
tenacity = ">=8.2.2"
tqdm = ">=4.65.0"
typer = ">=0.9.0"
unidecode = ">=1.3.6"

[tool.poetry.group.dev.dependencies]
black = { extras = ["jupyter"], version = ">=23.3.0" }
mypy = ">=1.1.1"
pandas-stubs = ">=2.0.1.230501"
pdbpp = ">=0.10.3"
ruff = ">=0.0.260"
types-tqdm = ">=4.65.0.1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
utmb-enrich = 'utmb_enrich.enrich_utmb:run'

[tool.black]
line-length = 100

[tool.mypy]
files = ["utmb_enrich"]
python_version = "3.11"
warn_unreachable = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
show_error_codes = true
ignore_missing_imports = false
exclude = [".git/", ".venv/", "__pycache__"]
follow_imports = "normal"
pretty = true
warn_incomplete_stub = true

[tool.ruff]
line-length = 100
target-version = "py311"

select = ["ALL"]

ignore = [
    "ANN101", # dont require self typing
    "ANN102", # dont require cls typing
    "COM812", # dont add trailing commas always
    "D100",   # dont force docstring presence
    "D101",   # dont force docstring presence
    "D102",   # dont force docstring presence
    "D103",   # dont force docstring presence
    "D104",   # dont force docstring presence
    "D105",   # dont force docstring presence
    "D106",   # dont force docstring presence
    "D203",   # conflict with other rule, this silences the warning
    "D213",   # conflict with other rule, this silences the warning
    "ERA",    # commenting out code is ok
    "G004",   # allow f-string in logging
    "S101",   # allow asserts
    "TRY400", # logger.error is not critical with loguru
]

unfixable = [
    "B007", # Don't automatically prefix loop control variable with "_" if unused
    "ERA",  # Don't delete commented code
    "F401", # Don't remove unused imports
    "F504", # Don't remove unused % format arguments
    "F522", # Don't remove unused .format arguments
    "F841", # Don't remove unused variables automatically
]

[tool.ruff.isort]
known-first-party = ["utmb-enrich"]
